♻️ Refactor : 기능에 알맞는 파일 구조 수정
